consolidate react-three-fiber stuff
