+++ src/app/visualization/three-fiber-components/Encampment.tsx
