+++ src/app/visualization/three-fiber-components/Ground.tsx
