+++ src/app/visualization/three-fiber-components/ResourceNode.tsx
